test: matlabs writeVideo; Idea: figure2im
Tetsting the capabilities of matlabs uncompressed 24 Bit video generation from single pictures/3D Matricies. Seems to work fine and should be able to display 1920x1080 videos in combination with WMP12 on the full screen.
Unsuccsessful tests: creating a matrix from for-loops and syms equations (too slow!); creating an image from a figure (figures were too small in the last test and it seemd to compress the image).
Implemented a colormap to discretize the figure into 'colorBits' (current variable name, may change) many Bits. 
Black Bars are probably needed to compensate the different aspect ratios of 16/9 and 1/0.63